Remove added route
