+++ packages/hub/services/queries/merchant-info.ts
@@ -36,6 +36,17 @@ export default class MerchantInfoQueries {
       [model.id, model.name, model.slug, model.color, model.textColor, model.ownerAddress]
     );
   }
+
+  async validateSlug(id: string) {
+    let db = await this.databaseManager.getClient();
+    let res = await db.query('SELECT slug from merchant_infos WHERE slug = $1', [id]);
+
+    if (res.rowCount === 0) {
+      return true;
+    } else {
+      return false;
+    }
+  }
 }
 
 declare module '@cardstack/hub/di/dependency-injection' {
